requirements.txt
python-telegram-bot>=20.7
sentence-transformers>=2.7.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
flask>=3.0.0
